Bug fixes for retrieving and converting terrain data files.
Multiple issues addresses: while loop allows for multiple iterations of
retrieval and conversion routine to completely get all A) .zip files B)
.hgt files, and C) .gz files. Tested using FRAPPE RF08.

In addition to terrain data issues, now the script exctracts GGLAT and
GGLON from the netCDF file as a fall-back in case LATC or LONC are not
available. Previously if LATC or LONC were missing, the values SFC_SRTM
values would be missing.